# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "QUIZ",
    "description": "QUIZ"
  },
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/addWallet": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "addWallet",
        "description": "addWallet",
        "operationId": "addWallet",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "$ref": "#/parameters/HeaderAuth"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/add_wallet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
    "/api/addGame": {
      "post": {
        "tags": [
          "ZZZ"
        ],
        "summary": "addGame",
        "description": "addGame",
        "operationId": "addGame",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "$ref": "#/parameters/HeaderAuth"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/add_game"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "add_wallet": {
      "type": "object",
      "required": [
        "balance"
        
      ],
      "properties": {
        "balance": {
          "type": "number",
          "example": "100"
        }
       
      }
    },
    "add_game": {
      "type": "object",
      "required": [
        "gameName",
        "isCompleted",
        "amount",
        "entrance"
        
      ],
      "properties": {
        "gameName": {
          "type": "string",
          "example": "Game Name"
        },
        "isCompleted": {
          "type": "boolean",
          "example": true
        },
        "amount": {
          "type": "number",
          "example": 500
        },
        "schedule": {
          "type": "number",
          "example": 1636690180000 
        },
        "entrance": {
          "type": "number",
          "example": 10
        }
       
      }
    },
...
    "responsesOk": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "msg": {
          "type": "string"
        },
        "data": {
          "type": "object"
        }
      }
    },
    "responsesError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "msg": {
          "type": "string"
        },
        "error": {
          "type": "object"
        }
      }
    },
...
  },
  "parameters": {
    "HeaderVersion": {
      "name": "X-QUIZ-Version",
      "in": "header",
      "description": "X-QUIZ-Version",
      "required": true,
      "type": "string",
      "default": "1.0.0"
    },
    "HeaderLanguage": {
      "name": "Accept-Language",
      "in": "header",
      "description": "Accept-Language",
      "required": true,
      "type": "string",
      "default": "en"
    },
    "HeaderPlatform": {
      "name": "X-QUIZ-Platform",
      "in": "header",
      "description": "android or ios or web",
      "required": true,
      "type": "string",
      "default": "android"
    },
    "HeaderAuth": {
      "name": "Authorization",
      "in": "header",
      "description": "Authorization",
      "required": false,
      "type": "string"
    },
...
  }
}